Rename resolvers for SitePreview
