+++ .env
@@ -42,7 +42,6 @@ API_PORT=4000
 API_URL=http://localhost:$API_PORT
 API_URL_INTERNAL=http://localhost:$API_PORT
 CORS_ALLOWED_ORIGINS="^http:\/\/localhost:\d+,^http://192.168.\d+.\d+:80[0-9]{2}"
-HMAC_SECRET=ACeLwPZhA3M9iyLAE946GXFydRz8YagKLH8c6sdwaJbe7WHQnmgEouebJZR4RxJJ
 
 # blob storage
 BLOB_STORAGE_DRIVER="file"
