+++ demo/api/src/config/environment-variables.ts
@@ -53,10 +53,6 @@ export class EnvironmentVariables {
 
     @IsString()
     @MinLength(16)
-    HMAC_SECRET: string;
-
-    @IsString()
-    @MinLength(16)
     DAM_SECRET: string;
 
     @IsString()
