+++ packages/api/cms-api/schema.gql
@@ -13,11 +13,6 @@ type PaginatedDependencies {
   totalCount: Int!
 }
 
-type Hmac {
-  timestamp: Float!
-  hash: String!
-}
-
 type ImageCropArea {
   focalPoint: FocalPoint!
   width: Float
@@ -257,6 +252,11 @@ type PaginatedPageTreeNodes {
   totalCount: Int!
 }
 
+type SitePreviewHash {
+  timestamp: Float!
+  hash: String!
+}
+
 type DamFolder {
   id: ID!
   name: String!
@@ -341,6 +341,8 @@ type Query {
   pageTreeNodeList(scope: PageTreeNodeScopeInput!, category: String): [PageTreeNode!]!
   paginatedPageTreeNodes(scope: PageTreeNodeScopeInput! = {}, category: String, sort: [PageTreeNodeSort!], offset: Int! = 0, limit: Int! = 25): PaginatedPageTreeNodes!
   pageTreeNodeSlugAvailable(scope: PageTreeNodeScopeInput!, parentId: ID, slug: String!): SlugAvailability!
+  getSitePreviewHash: SitePreviewHash!
+  validateSitePreviewHash(timestamp: Float!, hash: String!): Boolean!
   cronJobs: [CronJob!]!
   currentUser: CurrentUser!
   userPermissionsUserById(id: String!): User!
@@ -350,8 +352,6 @@ type Query {
   userPermissionsAvailablePermissions: [String!]!
   userPermissionsContentScopes(userId: String!, skipManual: Boolean): [JSONObject!]!
   userPermissionsAvailableContentScopes: [JSONObject!]!
-  hmacCreate: Hmac!
-  hmacValidate(timestamp: Float!, hash: String!): Boolean!
 }
 
 input RedirectScopeInput {
