+++ packages/api/cms-api/src/auth/resolver/auth.resolver.ts
@@ -1,7 +1,5 @@
 import { Type } from "@nestjs/common";
-import { Args, ArgsType, Context, Field, Mutation, ObjectType, Query, Resolver } from "@nestjs/graphql";
-import { IsNumber, IsString } from "class-validator";
-import crypto from "crypto";
+import { Context, Mutation, Query, Resolver } from "@nestjs/graphql";
 import { IncomingMessage } from "http";
 
 import { SkipBuild } from "../../builds/skip-build.decorator";
@@ -12,19 +10,6 @@ interface AuthResolverConfig {
     currentUser: Type<CurrentUserInterface>;
     endSessionEndpoint?: string;
     postLogoutRedirectUri?: string;
-    hmacSecret: string;
-}
-
-@ObjectType()
-@ArgsType()
-class Hmac {
-    @Field(() => Number)
-    @IsNumber()
-    timestamp: number;
-
-    @Field(() => String)
-    @IsString()
-    hash: string;
 }
 
 export function createAuthResolver(config: AuthResolverConfig): Type<unknown> {
@@ -49,32 +34,6 @@ export function createAuthResolver(config: AuthResolverConfig): Type<unknown> {
                 signOutUrl = url.toString();
             }
             return signOutUrl;
-        }
-
-        @Query(() => Hmac)
-        hmacCreate(): Hmac {
-            const timestamp = Math.floor(Date.now() / 1000);
-            return {
-                timestamp,
-                hash: this.createHash(timestamp),
-            };
-        }
-
-        @Query(() => Boolean)
-        hmacValidate(@Args() args: Hmac): boolean {
-            // Timestamp must be within the last 5 minutes
-            if (args.timestamp < Math.floor(Date.now() / 1000) - 60 * 5) {
-                return false;
-            }
-            return this.createHash(args.timestamp) === args.hash;
-        }
-
-        private createHash(timestamp: number): string {
-            if (!timestamp) throw new Error("Timestamp is required");
-            return crypto
-                .createHmac("sha256", config.hmacSecret)
-                .update(timestamp + config.hmacSecret)
-                .digest("hex");
         }
     }
     return AuthResolver;
